+++ src/main/java/com/telepathicgrunt/the_bumblezone/modcompat/ModChecker.java
@@ -22,6 +22,7 @@ public class ModChecker {
 	public static boolean twilightForestPresent = false;
 	public static boolean dragonEnchantsPresent = false;
 	public static boolean curiosPresent = false;
+	public static boolean goodallPresent = false;
 
 	/*
 	 * -- DO NOT TURN THE LAMBDAS INTO METHOD REFS. Method refs are not classloading safe. --
@@ -75,6 +76,9 @@ public class ModChecker {
 
 			modid = "curios";
 			loadupModCompat(modid, () -> CuriosCompat.setupCompat());
+
+			modid = "goodall";
+			loadupModCompat(modid, () -> GoodallCompat.setupCompat());
 		}
 		catch (Throwable e) {
 			printErrorToLogs("classloading " + modid + " and so, mod compat done afterwards broke");